RALPH: Review - extract shared repository context helpers
